veiculo-api/requirements.txt
django==3.0.5
djangorestframework==3.11.0
python-decouple==3.3